Add of test to sign out.  Test fails. Next step is to adjust layout template
